+++ network/p2p/internal/p2pfixtures/fixtures.go
@@ -1,0 +1,17 @@
+package p2pfixtures
+
+import (
+	"testing"
+
+	"github.com/onflow/flow-go/crypto"
+	"github.com/onflow/flow-go/utils/unittest"
+	"github.com/stretchr/testify/require"
+)
+
+// NetworkingKeyFixtures is a test helper that generates a ECDSA flow key pair.
+func NetworkingKeyFixtures(t *testing.T) crypto.PrivateKey {
+	seed := unittest.SeedFixture(48)
+	key, err := crypto.GeneratePrivateKey(crypto.ECDSASecp256k1, seed)
+	require.NoError(t, err)
+	return key
+}
