+++ network/p2p/translator/identity_provider_translator.go
@@ -1,4 +1,4 @@
-package p2p
+package translator
 
 import (
 	"fmt"
